Add kernel-package-override directory for make-kpkg
This addresses the initramfs build issue, where installing
the Grml kernel might result in a missing initrd because
kernel and initramfs-tools aren't installed in the correct
order. This is a make-kpkg-only issue, but since we still
build our kernels using kernel-package/make-kpkg we've
to apply this fix using the --overlay-dir option.